+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 # Change log
 
 ## [Unreleased]
+## Changed
+- rename gopen to ropen
+
+## Added
+- add encoding kwarg